better integration of two libraries
