+++ pyicloud/base.py
@@ -9,6 +9,7 @@ import sys
 import tempfile
 import os
 from re import match
+from uuid import uuid1 as generateClientID
 
 from pyicloud.exceptions import (
     PyiCloudFailedLoginException,
@@ -97,8 +98,8 @@ class PyiCloudSession(requests.Session):
         if not code and json.get('serverErrorCode'):
             code = json.get('serverErrorCode')
 
-        if reason:
-            self._raise_error(code, reason)
+        #if reason:
+        #    self._raise_error(code, reason)
 
         return response
 
@@ -106,13 +107,14 @@ class PyiCloudSession(requests.Session):
         if self.service.requires_2sa and \
                 reason == 'Missing X-APPLE-WEBAUTH-TOKEN cookie':
             raise PyiCloud2SARequiredError(response.url)
+
         if code == 'ZONE_NOT_FOUND' or code == 'AUTHENTICATION_FAILED':
             reason = 'Please log into https://icloud.com/ to manually ' \
                 'finish setting up your iCloud service'
             api_error = PyiCloudServiceNotActivatedErrror(reason, code)
             logger.error(api_error)
-
             raise(api_error)
+
         if code == 'ACCESS_DENIED':
             reason = reason + '.  Please wait a few minutes then try ' \
                 'again.  The remote servers might be trying to ' \
@@ -120,7 +122,7 @@ class PyiCloudSession(requests.Session):
 
         api_error = PyiCloudAPIResponseError(reason, code)
         logger.error(api_error)
-        raise api_error
+        raise(api_error)
 
 
 class PyiCloudService(object):
@@ -135,8 +137,8 @@ class PyiCloudService(object):
     """
 
     def __init__(
-        self, apple_id, password=None, cookie_directory=None, verify=True,
-        client_id=None
+            self, apple_id, password=None, cookie_directory=None, verify=True,
+            client_id=None
     ):
         if password is None:
             password = get_password_from_keyring(apple_id)
@@ -148,8 +150,12 @@ class PyiCloudService(object):
         self._password_filter = PyiCloudPasswordFilter(password)
         logger.addFilter(self._password_filter)
 
-        self._home_endpoint = 'https://www.icloud.com'
+        #self._home_endpoint =
+        self.user_agent = 'Opera/9.52 (X11; Linux i686; U; en)'
         self._setup_endpoint = 'https://setup.icloud.com/setup/ws/1'
+        self.referer = 'https://www.icloud.com'
+        self.origin = 'https://www.icloud.com'
+        self.response = None
 
         self._base_login_url = '%s/login' % self._setup_endpoint
 
@@ -166,9 +172,9 @@ class PyiCloudService(object):
         self.session = PyiCloudSession(self)
         self.session.verify = verify
         self.session.headers.update({
-            'Origin': self._home_endpoint,
-            'Referer': '%s/' % self._home_endpoint,
-            'User-Agent': 'Opera/9.52 (X11; Linux i686; U; en)'
+            'Origin': self.referer,
+            'Referer': '%s/' % self.referer,
+            'User-Agent': self.user_agent
         })
 
         cookiejar_path = self._get_cookiejar_path()
@@ -191,7 +197,18 @@ class PyiCloudService(object):
             'clientId': self.client_id,
         }
 
+        self.clientID = self.generateClientID()
+        self.setupiCloud = SetupiCloudService(self)
+        self.idmsaApple = IdmsaAppleService(self)
         self.authenticate()
+
+    def get_session_token(self):
+        self.clientID = self.generateClientID()
+        widgetKey = self.setupiCloud.requestAppleWidgetKey(self.clientID)
+        return self.idmsaApple.requestAppleSessionToken(self.user['apple_id'],
+                                                        self.user['password'],
+                                                        widgetKey
+                                                        )
 
     def authenticate(self):
         """
@@ -203,28 +220,16 @@ class PyiCloudService(object):
 
         data = dict(self.user)
 
-        # We authenticate every time, so "remember me" is not needed
+        sess_token = self.get_session_token()
 
-        myICloud = hack.PyiCloudService()
-        sess_token = myICloud.get_session_token (self.user['apple_id'],
-                                                 self.user['password']
-                                                 )
-                                                 
-        self.session.headers = {
-            'Origin': 'https://www.icloud.com',
-            'Referer': 'https://www.icloud.com/',
-            'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/70.0.3538.77 Safari/537.36',
-            'Content-Type': 'application/json',
-            'Accept': 'application/json, text/javascript, */*; q=0.01'
-        }
         data = {'accountCountryCode': "GBR",
-                    'extended_login': False,
-                    'dsWebAuthToken': sess_token
+                'extended_login': False,
+                'dsWebAuthToken': sess_token
                }
 
         try:
             req = self.session.post(
-                'https://setup.icloud.com/setup/ws/1/accountLogin',
+                self._setup_endpoint + '/accountLogin',
                 data=json.dumps(data)
             )
         except PyiCloudAPIResponseError as error:
@@ -244,6 +249,9 @@ class PyiCloudService(object):
 
         logger.info("Authentication completed successfully")
         logger.debug(self.params)
+
+    def generateClientID(self):
+        return str(generateClientID()).upper()
 
     def _get_cookiejar_path(self):
         # Get path for cookiejar file
@@ -376,3 +384,145 @@ class PyiCloudService(object):
 
     def __repr__(self):
         return '<%s>' % str(self)
+
+
+class HTTPService:
+    def __init__(self, session, response=None, origin=None, referer=None):
+        try:
+            self.session = session.session
+            self.response = session.response
+            self.origin = session.origin
+            self.referer = session.referer
+            self.user_agent = session.user_agent
+        except:
+            session = session
+            self.response = response
+            self.origin = origin
+            self.referer = referer
+            self.user_agent = "Python (X11; Linux x86_64)"
+
+
+class SetupiCloudService(HTTPService):
+    def __init__(self, session):
+        super(SetupiCloudService, self).__init__(session)
+        self.url = "https://setup.icloud.com/setup/ws/1"
+        self.urlKey = self.url + "/validate"
+        self.urlLogin = self.url + "/accountLogin"
+
+        self.appleWidgetKey = None
+        self.cookies = None
+        self.dsid = None
+
+    def requestAppleWidgetKey(self, clientID):
+        #self.urlBase + "/system/cloudos/16CHotfix21/en-us/javascript-packed.js"
+
+        self.session.headers.update(self.getRequestHeader())
+        self.response = self.session.get(self.urlKey, params=self.getQueryParameters(clientID))
+        try:
+            self.appleWidgetKey = self.findQyery(self.response.text, "widgetKey=")
+        except Exception as e:
+            raise Exception("requestAppletWidgetKey: Apple Widget Key query failed",
+                            self.urlKey, repr(e))
+        return self.appleWidgetKey
+
+    def requestCookies(self, appleSessionToken, clientID):
+        self.session.headers.update(self.getRequestHeader())
+        self.response = self.session.post(self.urlLogin,
+                                          self.getLoginRequestPayload(appleSessionToken),
+                                          params=self.getQueryParameters(clientID))
+        try:
+            self.cookies = self.response.headers["Set-Cookie"]
+        except Exception as e:
+            raise Exception("requestCookies: Cookies query failed",
+                            self.urlLogin, repr(e))
+        try:
+            self.dsid = self.response.json()["dsInfo"]["dsid"]
+        except Exception as e:
+            raise Exception("requestCookies: dsid query failed",
+                            self.urlLogin, repr(e))
+        return self.cookies, self.dsid
+
+    def findQyery(self, data, query):
+        response = ''
+        foundAt = data.find(query)
+        if foundAt == -1:
+            raise Exception("findQyery: " + query + " could not be found in data")
+        foundAt += len(query)
+        char = data[foundAt]
+        while char.isalnum():
+            response += char
+            foundAt += 1
+            char = data[foundAt]
+        return response
+
+
+    def getRequestHeader(self):
+        header = {
+            "Accept": "*/*",
+            "Connection": "keep-alive",
+            "Content-Type": "text/plain",
+            "User-Agent": self.user_agent,
+            "Origin": self.origin,
+            "Referer": self.referer,
+            }
+        return header
+
+    def getQueryParameters(self, clientID):
+        if not clientID:
+            raise NameError("getQueryParameters: clientID not found")
+        return {
+            "clientBuildNumber": "16CHotfix21",
+            "clientID": clientID,
+            "clientMasteringNumber": "16CHotfix21",
+            }
+
+    def getLoginRequestPayload(self, appleSessionToken):
+        if not appleSessionToken:
+            raise NameError("getLoginRequestPayload: X-Apple-ID-Session-Id not found")
+        return json({
+            "dsWebAuthToken": appleSessionToken,
+            "extended_login": False,
+            })
+
+class IdmsaAppleService(HTTPService):
+    def __init__(self, session):
+        super(IdmsaAppleService, self).__init__(session)
+        self.url = "https://idmsa.apple.com"
+        self.urlAuth = self.url + "/appleauth/auth/signin?widgetKey="
+
+        self.appleSessionToken = None
+
+    def requestAppleSessionToken(self, user, password, appleWidgetKey):
+        self.session.headers.update(self.getRequestHeader(appleWidgetKey))
+        self.response = self.session.post(self.urlAuth + appleWidgetKey,
+                                          self.getRequestPayload(user, password))
+        try:
+            self.appleSessionToken = self.response.headers["X-Apple-Session-Token"]
+        except Exception as e:
+            raise Exception("requestAppleSessionToken: Apple Session Token query failed",
+                            self.urlAuth, repr(e))
+        return self.appleSessionToken
+
+    def getRequestHeader(self, appleWidgetKey):
+        if not appleWidgetKey:
+            raise NameError("getRequestHeader: clientID not found")
+        return {
+            "Accept": "application/json, text/javascript",
+            "Content-Type": "application/json",
+            "User-Agent": self.user_agent,
+            "X-Apple-Widget-Key": appleWidgetKey,
+            "X-Requested-With": "XMLHttpRequest",
+            "Origin": self.origin,
+            "Referer": self.referer,
+            }
+
+    def getRequestPayload(self, user, password):
+        if not user:
+            raise NameError("getAuthenticationRequestPayload: user not found")
+        if not password:
+            raise NameError("getAuthenticationRequestPayload: password not found")
+        return json.dumps({
+            "accountName": user,
+            "password": password,
+            "rememberMe": False,
+            })
